Stop manipulating the returned file URL in RCS
I thought it would be a good idea to reformat the file
URL as it came from canvas, but it turns out the URL
needs to change based on how the URL is being used in
the RCE (i.e. as a link, or as an embedded image).
Let's let the RCE decide.

refs LA-585
flag=none

test plan:
  - query the rcs /api/documents?contextType=course&contextId=:id
  > expect the URLs to be the /files/:id urls directly from canvas
    for example /files/17/download?download_frd=1

Change-Id: I554805a6752def915c187e816b8ed0ec34199dcd
Reviewed-on: https://gerrit.instructure.com/c/canvas-rce-api/+/225943
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jon Willesen <jonw+gerrit@instructure.com>
QA-Review: Jon Willesen <jonw+gerrit@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>